+++ workmanager/lib/src/workmanager_impl.dart
@@ -220,7 +220,7 @@ class Workmanager {
     Map<String, dynamic>? inputData,
     Duration? initialDelay,
     Constraints? constraints,
-    ExistingWorkPolicy? existingWorkPolicy,
+    ExistingPeriodicWorkPolicy? existingWorkPolicy,
     BackoffPolicy? backoffPolicy,
     Duration? backoffPolicyDelay,
     String? tag,